Added Switch for llama-cpp & ollama with LOCALLLAMA Provider
